pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mednexus"
version = "0.1.0"
description = "MedNexus – Multi-Agent Healthcare Orchestration Platform"

requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "MedNexus Team" }]

dependencies = [
    # ── Azure SDK ────────────────────────────────────────────
    "azure-identity>=1.16",
    "azure-cosmos>=4.7",
    "azure-search-documents>=11.6",
    "azure-ai-inference>=1.0",
    "azure-storage-blob>=12.20",
    "azure-functions>=1.20",
    # ── AI / LLM ─────────────────────────────────────────────
    "openai>=1.40",
    "semantic-kernel>=1.5",
    # ── Web / API ────────────────────────────────────────────
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=13.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    # ── Data & Validation ────────────────────────────────────
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    # ── Observability ────────────────────────────────────────
    "azure-monitor-opentelemetry>=1.6",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "structlog>=24.2",
    # ── MCP Protocol ─────────────────────────────────────────
    "mcp>=1.0",
    # ── Auth ─────────────────────────────────────────────────
    "PyJWT>=2.8",
    # ── Utilities ────────────────────────────────────────────
    "python-dotenv>=1.0",
    "aiofiles>=24.1",
    "pillow>=10.4",
    "PyPDF2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
